Fixed overflow problem with LCG when a is too big. Added LSB2 algorithm which combines lsbk and lsbp

+++ src/steg/lcg.h
@@ -8,7 +8,7 @@ class LCG {
     int m, c;
     int trueM;
     int a;
-    int seed;
+    long long seed;
     int gcd(int u, int v);
   public:
     LCG() {};
